// apps/web/src/api/model/createCardProgramDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { CardProgramAuthorizationDto } from './cardProgramAuthorizationDto';
import type { CardProgramDistributionDto } from './cardProgramDistributionDto';
import type { CardProgramPersonalizationDto } from './cardProgramPersonalizationDto';
import type { CreateCardProgramDtoCurrency } from './createCardProgramDtoCurrency';
import type { CreateCardProgramDtoNetwork } from './createCardProgramDtoNetwork';
import type { CreateCardProgramDtoType } from './createCardProgramDtoType';

export interface CreateCardProgramDto {
  name: string;
  description?: string;
  network: CreateCardProgramDtoNetwork;
  type: CreateCardProgramDtoType;
  quantity?: number;
  /** Card Bin ID */
  bin?: string;
  currency: CreateCardProgramDtoCurrency;
  distribution?: CardProgramDistributionDto;
  personalization?: CardProgramPersonalizationDto;
  authorization: CardProgramAuthorizationDto;
}