// apps/web/src/api/model/patchCardProgramDraftDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Allawee Infrastructure Backend APIs
 * The Allawee Infrastructure APIs description
 * OpenAPI spec version: 0.0.1
 */
import type { CardProgramAuthorizationDto } from './cardProgramAuthorizationDto';
import type { CardProgramDistributionDto } from './cardProgramDistributionDto';
import type { CardProgramPersonalizationDto } from './cardProgramPersonalizationDto';
import type { PatchCardProgramDraftDtoCurrency } from './patchCardProgramDraftDtoCurrency';
import type { PatchCardProgramDraftDtoNetwork } from './patchCardProgramDraftDtoNetwork';

export interface PatchCardProgramDraftDto {
  name?: string;
  description?: string;
  network?: PatchCardProgramDraftDtoNetwork;
  quantity?: number;
  /** Card Bin ID */
  bin?: string;
  currency?: PatchCardProgramDraftDtoCurrency;
  distribution?: CardProgramDistributionDto;
  personalization?: CardProgramPersonalizationDto;
  authorization?: CardProgramAuthorizationDto;
}